requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.2.3 
langchain_pinecone
flask
pypdf
langchain-community
python-dotenv
-e .